Relocate TasksCompletionSchedulerProperties to new package and update dependencies
